Rules (and a new Rule class) to implement alternation rules on morpheme sequences
Output of segmentation is a sequence of morphemes separated by '-', with the stem surrounded by braces.
To spell this, there are now rules based on regexs, implementing the rules in the FSTs for each POS.
Language has a method join_segments() to do this, given a POS label for a set of rules.