Fidel analysis: improvements in roots.py and tmp.py
Weak subclasses now add missing templates from strong defaults unless a feature set is explicitly excluded (with '!' in a subclass line in the .tmp file).
For roots with multiple "weak" consonants, priority is for the earliest one. For example, <' w q> belongs to '1 rather than 2w.
This needs to be fixed later on for verbs belonging to multiple weak classes like አየ and አማ.